Moves Combine helpers to a new file + reorganizes and marks the two seperate implementations of the composition.
