+++ FeedApplication/FeedApplication/SceneDelegate.swift
@@ -6,7 +6,6 @@
 //
 
 import UIKit
-import Combine
 import CoreData
 import FeedCoreModule
 import FeediOS
@@ -40,38 +39,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window = UIWindow(windowScene: scene)
         configureWindow()
     }
-      
-//    func configureWindow() {
-//        let url = URL(string: "https://ile-api.essentialdeveloper.com/essential-feed/v1/feed")!
-//        let client = makeRemoteClient()
-//        let remoteFeedLoader = RemoteFeedLoader(url: url, client: client)
-//        let remoteImageLoader = RemoteImageLoader(client: client)
-//        let localImageLoader = LocalImageLoader(store: store)
-        
-//        window?.rootViewController = UINavigationController(
-//            rootViewController: FeedUIComposer.feedControllerComposedWith(
-//        feedLoader: FeedLoaderWithFallbackComposite(
-//            primary: FeedLoaderCacheDecorator(
-//                decoratee: remoteFeedLoader,
-//                cache: localFeedLoader),
-//            fallback: localFeedLoader),
-//        imageLoader: ImageLoaderWithFallbackComposite(
-//            primary: localImageLoader,
-//            fallback: ImageLoaderCacheDecorator(
-//                decoratee: remoteImageLoader,
-//                cache: localImageLoader))))
-//    }
-
-    // MARK: - Combine Alternative to composition
-
-    func configureWindow() {
-        window?.rootViewController = UINavigationController(
-            rootViewController: FeedUIComposer.feedControllerComposedWith(
-        feedLoader: makeRemoteFeedLoaderWithLocalFallback,
-        imageLoader: makeLocalImageLoaderWithRemoteFallback))
-        
-        window?.makeKeyAndVisible()
-    }
     
     func sceneWillResignActive(_ scene: UIScene) {
         localFeedLoader.validateCache { _ in }
@@ -79,6 +46,17 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     
     func makeRemoteClient() -> HTTPClient {
         return httpClient
+    }
+    
+    // MARK: - Composition with Combine Framework + universal abstractions
+    
+    func configureWindow() {
+        window?.rootViewController = UINavigationController(
+            rootViewController: FeedUIComposer.feedControllerComposedWith(
+        feedLoader: makeRemoteFeedLoaderWithLocalFallback,
+        imageLoader: makeLocalImageLoaderWithRemoteFallback))
+        
+        window?.makeKeyAndVisible()
     }
     
     private func makeRemoteFeedLoaderWithLocalFallback() -> FeedLoader.Publisher {
@@ -103,104 +81,29 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
                     .cacheing(to: localImageLoader, using: url)
             })
     }
-}
-
-public extension FeedLoader {
-    typealias Publisher = AnyPublisher<[FeedImage], Error>
-    func loadPublisher() -> Publisher {
-        Deferred {
-            Future(self.load)
-        }
-        .eraseToAnyPublisher()
-    }
-}
-
-extension Publisher where Output == [FeedImage] {
-    func cacheing(to cache: FeedCache) -> AnyPublisher<Output, Failure> {
-        handleEvents(receiveOutput: cache.saveIgnoringResult).eraseToAnyPublisher()
-    }
-}
-
-private extension FeedCache {
-    func saveIgnoringResult(_ feed: [FeedImage]) {
-        save(feed) { _ in }
-    }
-}
-
-public extension ImageLoader {
-    typealias Publisher = AnyPublisher<Data, Error>
     
-    func loadImagePublisher(with url: URL) -> Publisher {
-        var task: ImageLoaderTask?
-        
-        return Deferred {
-            Future { completion in
-                task = self.loadImageData(from: url, completion: completion)
-            }
-        }
-        .handleEvents(receiveCancel: { task?.cancel() })
-        .eraseToAnyPublisher()
-    }
-}
-
-extension Publisher where Output == Data {
-    func cacheing(to cache: ImageCache, using url: URL) -> AnyPublisher<Output, Failure> {
-        handleEvents(receiveOutput: { data in
-            cache.saveIgnoringResult(data, for: url)
-        }).eraseToAnyPublisher()
-    }
-}
-
-private extension ImageCache {
-    func saveIgnoringResult(_ data: Data, for url: URL) {
-        save(data, for: url) { _ in }
-    }
-}
-
-extension Publisher {
-    func fallback(to fallbackPublisher: @escaping () -> AnyPublisher<Output, Failure>) -> AnyPublisher<Output, Failure> {
-        self.catch { _ in fallbackPublisher() }.eraseToAnyPublisher()
-    }
-}
-
-extension Publisher {
-    func dispatchOnMainQueue() -> AnyPublisher<Output, Failure> {
-        receive(on: DispatchQueue.immediateWhenOnMainQueueScheduler).eraseToAnyPublisher()
-    }
-}
-
-extension DispatchQueue {
     
-    static var immediateWhenOnMainQueueScheduler: ImmediateWhenOnMainQueueScheduler {
-        ImmediateWhenOnMainQueueScheduler()
-    }
+    // MARK: - Composition with Design Patterns
     
-    struct ImmediateWhenOnMainQueueScheduler: Scheduler {
-        typealias SchedulerTimeType = DispatchQueue.SchedulerTimeType
-        typealias SchedulerOptions = DispatchQueue.SchedulerOptions
+//    func configureWindow() {
+//        let url = URL(string: "https://ile-api.essentialdeveloper.com/essential-feed/v1/feed")!
+//        let client = makeRemoteClient()
+//        let remoteFeedLoader = RemoteFeedLoader(url: url, client: client)
+//        let remoteImageLoader = RemoteImageLoader(client: client)
+//        let localImageLoader = LocalImageLoader(store: store)
         
-        var now: Self.SchedulerTimeType {
-            DispatchQueue.main.now
-        }
-
-        var minimumTolerance: Self.SchedulerTimeType.Stride {
-            DispatchQueue.main.minimumTolerance
-        }
-
-        func schedule(options: Self.SchedulerOptions?, _ action: @escaping () -> Void) {
-            guard Thread.isMainThread else {
-                DispatchQueue.main.schedule(options: options, action)
-                return
-            }
-            action()
-        }
-        
-        func schedule(after date: Self.SchedulerTimeType, tolerance: Self.SchedulerTimeType.Stride, options: Self.SchedulerOptions?, _ action: @escaping () -> Void) {
-                DispatchQueue.main.schedule(after: date, tolerance: tolerance, options: options, action)
-        }
-
-        func schedule(after date: Self.SchedulerTimeType, interval: Self.SchedulerTimeType.Stride, tolerance: Self.SchedulerTimeType.Stride, options: Self.SchedulerOptions?, _ action: @escaping () -> Void) -> Cancellable {
-            DispatchQueue.main.schedule(after: date, interval: interval, tolerance: tolerance, options: options, action)
-        }
-    }
+//        window?.rootViewController = UINavigationController(
+//            rootViewController: FeedUIComposer.feedControllerComposedWith(
+//        feedLoader: FeedLoaderWithFallbackComposite(
+//            primary: FeedLoaderCacheDecorator(
+//                decoratee: remoteFeedLoader,
+//                cache: localFeedLoader),
+//            fallback: localFeedLoader),
+//        imageLoader: ImageLoaderWithFallbackComposite(
+//            primary: localImageLoader,
+//            fallback: ImageLoaderCacheDecorator(
+//                decoratee: remoteImageLoader,
+//                cache: localImageLoader))))
+//    }
+    
 }
